requirements.txt
efficientnet==1.1.1
fire==0.4.0
geojson==2.5.0
grpcio==1.42.0
h5py==2.10.0
imageio==2.19.1
imutils==0.5.4
ipython-genutils==0.2.0
ipywidgets==7.7.1
jupyterlab-widgets==1.1.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.9.0
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
openpyxl==3.0.10
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyarrow==8.0.0
pydeck==0.7.1
Pympler==1.0.1
PyPDF2==1.26.0
PyWavelets==1.3.0
requests
rich==12.5.1
scikit-image==0.19.2
scikit-learn==1.0.1
scipy==1.7.3
semver==2.13.0
setuptools-scm==6.3.2
simplejson==3.17.6
sklearn==0.0
smmap==5.0.0
tabula-py==2.4.0
tdqm==0.0.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
toolz==0.12.0
tqdm==4.62.3
validators==0.20.0
watchdog==2.1.9
webencodings==0.5.1
Werkzeug==2.0.2
